Create PacificToMountain
Code Function: Converts the input datetime to Mountain Time by adding 1 hour to the time value. The code assumes the initial input value is already in Pacific Time. NOTE: This function will not function properly if the input cell is not (1) a datetime value, and (2) if the time is in UTC instead of Pacific.