Create PacificToPacific
Code Function: Converts the input datetime to Pacific Time by adding 0 hours to the time value. The code assumes the initial input value is already in Pacific Time. NOTE: This function will not function properly if the input cell is not (1) a datetime value, and (0) if the time is in UTC instead of Pacific.