Create compileList.gs
Author: Nick Taylor

Purpose: Requires user to provide the URL to the CAR (Client Account Review) so that Client names can be extracted from each tab into a transposed list. Setting up the user to continue through remaining steps to populate the highest assistance level per Care Pro.

Script Callout: The URL is stored using the Properties Service for use in the Populate Assistance Levels script. This is done over variable storage so that the URL can be called in other script executions.