Create populateAssistanceLevels3
Author: Nick Taylor

Purpose: In the provided CAR (Client Account Review) workbook, the script references each Client Name tab to find the <Assistance Level> for each Client by finding the header "Assistance Level", and selecting the value directly to the right of the header. The value is then populated on the Compiled List tab in the TEMPLATE workbook, under the "Week 4" column, Column D.

Script Callout: The method is performed this way to account for any changes in formatting. Normally, the Assistance Level can be found in Cell D23, but if the formatting changes, the wrong value could be populated, thus erroring out the script. This method accounts for that by searching for the header and populating the value next to it.

Script Callout: The IMPL Franchise Tracker sometimes requires Specialists to run the Assistance Levels in 3 phases: Initial, Week 3, and Week 4. This script allows for a snapshot view of Assistance Levels per Client, without overwriting previous data.